Separates configuration for tasks and module filtering
