+++ talaiot/src/main/kotlin/com/cdsap/talaiot/configuration/FilterConfiguration.kt
@@ -1,15 +1,32 @@
 package com.cdsap.talaiot.configuration
 
-import java.util.regex.Pattern
+import com.cdsap.talaiot.configuration.filter.ModuleFilterConfiguration
+import com.cdsap.talaiot.configuration.filter.TaskFilterConfiguration
+import groovy.lang.Closure
 
 class FilterConfiguration {
-    var includes: Array<String> = emptyArray()
-    var excludes: Array<String> = emptyArray()
+    var tasks: TaskFilterConfiguration? = null
+    var modules: ModuleFilterConfiguration? = null
 
 
-    private fun shouldInclude(moduleName: String) =
-            includes.find { moduleName.matches(Regex.fromLiteral(Pattern.quote(it))) } != null
+    fun tasks(configuration: TaskFilterConfiguration.() -> Unit) {
+        tasks = TaskFilterConfiguration().also(configuration)
+    }
 
-    private fun shouldExclude(taskName: String) =
-            excludes.find { taskName.matches(Regex.fromLiteral(Pattern.quote(it))) } != null
+    fun tasks(closure: Closure<*>) {
+        tasks = TaskFilterConfiguration()
+        closure.delegate = tasks
+        closure.call()
+    }
+
+
+    fun modules(configuration: ModuleFilterConfiguration.() -> Unit) {
+        modules = ModuleFilterConfiguration().also(configuration)
+    }
+
+    fun modules(closure: Closure<*>) {
+        modules = ModuleFilterConfiguration()
+        closure.delegate = modules
+        closure.call()
+    }
 }
