+++ talaiot/src/main/kotlin/com/cdsap/talaiot/configuration/filter/TaskFilterConfiguration.kt
@@ -1,0 +1,6 @@
+package com.cdsap.talaiot.configuration.filter
+
+class TaskFilterConfiguration {
+    var includes: Array<String> = emptyArray()
+    var excludes: Array<String> = emptyArray()
+}
